Add files via upload
(a). On the Create notebook instance page, in the Notebook instance setting box, fill the following fields:

1.For Notebook instance name, type SageMaker-Tutorial.
2.For Notebook instance type, choose ml.t2.medium.
3.For Elastic inference, keep the default selection of none.
4.For Platform identifier, keep the default selection.

(b). In the Permissions and encryption section, in the IAM role drop down, and choose Create a new role. In the Create an IAM role dialog box, select Any S3 bucket, and choose Create role.

Note: If you already have a bucket that you’d like to use instead, choose Specific S3 buckets and specify the bucket name.